\name{spcma-package}
\alias{spcma-package}
\docType{package}
\title{
\packageTitle{spcma}
}
\description{
\packageDescription{spcma}
}
\details{
\tabular{ll}{
	Package: \tab spcma \cr
	Type: \tab Package \cr
	Version: \tab 1.0 \cr
	Date: \tab 2018-06-08 \cr
	License: \tab GPL (>=2) \cr
}
}
\author{
\packageAuthor{spcma}

Maintainer: \packageMaintainer{spcma}
}
\keyword{ package }
